[refactor] 상품 세부 페이지 컴포넌트 분리
